remove ascii_part != NULL check (no one is doing that), remove extra white spaces in fetch.c, replace http_code by response_code, response_code should be long instead of int

+++ src/cloudpinyin.h
@@ -60,7 +60,7 @@ typedef struct _CurlQueue
     struct _CurlQueue* next;
     CloudPinyinRequestType type;
     int curl_result;
-    int http_code;
+    long http_code;
     char* str;
     char* pinyin;
     size_t size;
